requirements.txt
pystray
pynput
sounddevice
scipy
numpy
openai
openai-whisper
python-dotenv
soundfile 